fix(features): harden windowed multimodal feature extraction layer
- Redesign feature coverage semantics in FeatureCoverage and FeatureQuality based on usable window duration
- Implement EventCursor for monotonic index range lookup across sliding windows (O(N + W) time)
- Implement Terminating R-Peak Boundary Convention in cardiac_features for boundary-crossing RR intervals
- Implement time_range_to_sample_range for strict physical-to-sample half-open interval conversions
- Enforce signal dimension equality (tonic.len() == phasic.len()) in eda_features with SignalError::DimensionMismatch
- Document explicit population standard deviation formula for SDNN and field alias rr_std_ms
- Implement PrecomputedCoupling for recording-level coupling computation and fast window aggregation
- Introduce TimedEvents and TimedSignal containers in MultimodalInput with fallible builder methods
- Add extract_features_naive reference implementation for dual-implementation oracle equivalence testing
- Add test coverage in tests/features_tests.rs (coverage semantics, SDNN known values, boundary RR intervals, time range conversion, oracle parity, 100k scaling)
- Add 1M event scaling benchmarks in benches/benchmarks.rs
- Update docs/architecture.md Section 6 and README.md

+++ tests/multimodal_tests.rs
@@ -328,8 +328,8 @@ fn test_multi_sampling_rate_timing_invariance() {
     let sampling_rates = [32.0, 64.0, 100.0, 250.0, 500.0, 1000.0];
 
     for &fs in &sampling_rates {
-        let ecg_idx = (1.0 * fs as f64).round() as usize;
-        let ppg_idx = (1.2 * fs as f64).round() as usize;
+        let ecg_idx = (1.0f64 * fs).round() as usize;
+        let ppg_idx = (1.2f64 * fs).round() as usize;
 
         let matches = ecg_ppg_timing(&[ecg_idx], fs, 0.0, &[ppg_idx], fs, 0.0).unwrap();
         assert_eq!(matches.len(), 1);
